// mobile/src/common/enums/tracked-event-name.enum.ts
export enum TrackedEventName {
  // Access Request
  APPROVE_ACCESS_REQUEST = 'APPROVE_ACCESS_REQUEST',
  REJECT_ACCESS_REQUEST = 'REJECT_ACCESS_REQUEST',

  // Activity Log
  CHANGE_ACTIVITY_LOG_STATUS = 'CHANGE_ACTIVITY_LOG_STATUS',

  // Documents
  CREATE_CONTRACT = 'CREATE_CONTRACT',
  APPROVE_CONTRACT = 'APPROVE_CONTRACT',
  REJECT_CONTRACT = 'REJECT_CONTRACT',

  // Document Contract - new implementation
  CREATE_DOCUMENT_CONTRACT = 'CREATE_DOCUMENT_CONTRACT',
  VALIDATE_DOCUMENT_CONTRACT = 'VALIDATE_DOCUMENT_CONTRACT',
  SIGN_DOCUMENT_CONTRACT_BY_NGO = 'SIGN_DOCUMENT_CONTRACT_BY_NGO',
  SIGN_DOCUMENT_CONTRACT_BY_VOLUNTEER = 'SIGN_DOCUMENT_CONTRACT_BY_VOLUNTEER',
  REJECT_DOCUMENT_CONTRACT_BY_NGO = 'REJECT_DOCUMENT_CONTRACT_BY_NGO',
  REJECT_DOCUMENT_CONTRACT_BY_VOLUNTEER = 'REJECT_DOCUMENT_CONTRACT_BY_VOLUNTEER',
  DELETE_DOCUMENT_CONTRACT = 'DELETE_DOCUMENT_CONTRACT',
}